+++ src/server/mod.rs
@@ -5,5 +5,4 @@
 pub mod api;
 pub mod app;
 pub mod errors;
-pub mod git;
 pub mod tracing;
